Refactor, implement CachedResolver.
