--- lib/example-did-driver.js
@@ -1,10 +1,0 @@
-/*!
- * Copyright (c) 2021 Digital Bazaar, Inc. All rights reserved.
- */
-// This would handle 'did:example:... DIDs
-export const method = 'example';
-
-// export async function generate({keyMap, cryptoLd} = {}) {}
-
-export async function get() {}
-
