+++ lib/main.js
@@ -8,6 +8,6 @@ export {
   parseDid
 } from './did-io.js';
 
-export {MultiResolver} from './MultiResolver.js';
+export {CachedResolver} from './CachedResolver.js';
 
 export {VERIFICATION_RELATIONSHIPS} from './constants.js';
