+++ package.json
@@ -41,6 +41,7 @@
     "lib/*.js"
   ],
   "dependencies": {
+    "@digitalbazaar/lru-memoize": "^2.0.0",
     "did-context": "^2.0.0",
     "esm": "^3.2.25"
   },
